requirements.txt
streamlit
git+https://github.com/streamlit/gsheets-connection
pandasql  # this is for example/st_app.py only
st-gsheets-connection
langchain-groq
langchain
langchain-openai
langchain-community
faiss-cpu

